restructured project to be more tidy. added model files to wrap meshes and separated between triangle mesh and scene entity. removed warnings.
