+++ programming_along/scene/Camera.cpp
